Fix type mismatch in PlayerDataCollector skill experience collection
Cast getSkillExperience() return value from int to long to match the
Map<String, Long> type declaration. This resolves the compilation error
on line 68.

+++ osrs-ai-assistant/src/main/java/net/runelite/client/plugins/aiassistant/data/PlayerDataCollector.java
@@ -65,7 +65,7 @@ public class PlayerDataCollector
 		{
 			String skillName = skill.getName();
 			skillLevels.put(skillName, client.getRealSkillLevel(skill));
-			skillExperience.put(skillName, client.getSkillExperience(skill));
+			skillExperience.put(skillName, (long) client.getSkillExperience(skill));
 		}
 		builder.skillLevels(skillLevels);
 		builder.skillExperience(skillExperience);
